tests: Extract compat tests and don't use pytester in fixture tests
